requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.41.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
datasets==2.19.1
psutil==5.9.8
pandas==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
